SocialNetwork_DjangoProject/social_network/requirements.txt
asgiref==3.7.2
distlib==0.3.7
Django==4.2.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
package-name==0.1
platformdirs==3.11.0
PyJWT==2.9.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
virtualenv==20.24.6
